+++ integration-tests/functions/src/api/ignore/foo.d.ts
@@ -1,0 +1,1 @@
+declare const test = 'test'