+++ packages/gatsby-cli/src/reporter/redux/diagnostics.ts
@@ -39,7 +39,7 @@ const TEN_MINUTES = 1000 * 60 * 10
 const TEN_SECONDS = 1000 * 10
 
 export function createStructuredLoggingDiagnosticsMiddleware(
-  store: GatsbyCLIStore
+  getStore: () => GatsbyCLIStore
 ): DiagnosticsMiddleware {
   const stuckStatusDiagnosticTimeoutDelay = calculateTimeoutDelay(
     process.env.GATSBY_DIAGNOSTIC_STUCK_STATUS_TIMEOUT,
@@ -67,7 +67,7 @@ export function createStructuredLoggingDiagnosticsMiddleware(
   function inProgressActivities(): Array<
     IActivity & { diagnostics_elapsed_seconds?: string }
   > {
-    const { activities } = store.getState().logs
+    const { activities } = getStore().getState().logs
     return Object.values(activities)
       .filter(activity => isActivityInProgress(activity.status))
       .map(activity => {
@@ -96,7 +96,7 @@ export function createStructuredLoggingDiagnosticsMiddleware(
     // ignore diagnostic logs, otherwise diagnostic message itself will reset
     // the timers
     if (!displayingStuckStatusDiagnosticWarning) {
-      const currentStatus = store.getState().logs.status
+      const currentStatus = getStore().getState().logs.status
 
       if (!reporter) {
         // yuck, we have situation of circular dependencies here
@@ -122,7 +122,7 @@ export function createStructuredLoggingDiagnosticsMiddleware(
                 null,
                 2
               )}\n\nCurrently Gatsby is in: "${
-                store.getState().logs.status
+                getStore().getState().logs.status
               }" state.${
                 activitiesDiagnosticsMessage.length > 0
                   ? `\n\nActivities preventing Gatsby from transitioning to idle state:\n\n${activitiesDiagnosticsMessage}`
@@ -139,7 +139,7 @@ export function createStructuredLoggingDiagnosticsMiddleware(
               displayingStuckStatusDiagnosticWarning = true
               reporter.warn(
                 `This is just diagnostic information (enabled by GATSBY_DIAGNOSTIC_STUCK_STATUS_TIMEOUT):\n\nGatsby is in "${
-                  store.getState().logs.status
+                  getStore().getState().logs.status
                 }" state without any updates for ${(
                   stuckStatusDiagnosticTimeoutDelay / 1000
                 ).toFixed(
@@ -175,7 +175,7 @@ export function createStructuredLoggingDiagnosticsMiddleware(
                 id: `11701`,
                 context: {
                   activities: inProgressActivities(),
-                  status: store.getState().logs.status,
+                  status: getStore().getState().logs.status,
                   stuckStatusDiagnosticMessage:
                     generateStuckStatusDiagnosticMessage(),
                   stuckStatusWatchdogTimeoutDelay,