+++ packages/gatsby/src/bootstrap/load-plugins/index.ts
@@ -22,9 +22,9 @@ import {
 } from "./types"
 import { IPluginRefObject, PluginRef } from "gatsby-plugin-utils/dist/types"
 
-const getAPI = (
-  api: { [exportType in ExportType]: { [api: string]: boolean } }
-): ICurrentAPIs =>
+const getAPI = (api: {
+  [exportType in ExportType]: { [api: string]: boolean }
+}): ICurrentAPIs =>
   _.keys(api).reduce<Partial<ICurrentAPIs>>((merged, key) => {
     merged[key] = _.keys(api[key])
     return merged
@@ -51,6 +51,7 @@ const flattenPlugins = (plugins: Array<IPluginInfo>): Array<IPluginInfo> => {
             roots = roots.map(root => root[segment])
           }
         }
+        roots = roots.flat()
 
         roots.forEach(subPlugin => {
           flattened.push(subPlugin)