+++ packages/gatsby/src/schema/graphql-engine/bundle-webpack.ts
@@ -5,7 +5,7 @@ import * as fs from "fs-extra"
 import webpack from "webpack"
 import { printQueryEnginePlugins } from "./print-plugins"
 import mod from "module"
-import { WebpackLoggingPlugin } from "../../utils/webpack/webpack-logging"
+import { WebpackLoggingPlugin } from "../../utils/webpack/plugins/webpack-logging"
 import reporter from "gatsby-cli/lib/reporter"
 
 type Reporter = typeof reporter
@@ -13,6 +13,12 @@ type Reporter = typeof reporter
 const extensions = [`.mjs`, `.js`, `.json`, `.node`, `.ts`, `.tsx`]
 
 const outputDir = path.join(process.cwd(), `.cache`, `query-engine`)
+const cacheLocation = path.join(
+  process.cwd(),
+  `.cache`,
+  `webpack`,
+  `query-engine`
+)
 
 export async function createGraphqlEngineBundle(
   rootDir: string,
@@ -38,6 +44,14 @@ export async function createGraphqlEngineBundle(
     target: `node`,
     externalsPresets: {
       node: false,
+    },
+    cache: {
+      type: `filesystem`,
+      name: `graphql-engine`,
+      cacheLocation,
+      buildDependencies: {
+        config: [__filename],
+      },
     },
     // those are required in some runtime paths, but we don't need them
     externals: [
@@ -89,6 +103,10 @@ export async function createGraphqlEngineBundle(
       extensions,
       alias: {
         ".cache": process.cwd() + `/.cache/`,
+
+        [require.resolve(`gatsby-cli/lib/reporter/loggers/ink/index.js`)]:
+          false,
+        inquirer: false,
       },
     },
     plugins: [
@@ -97,6 +115,7 @@ export async function createGraphqlEngineBundle(
         "process.env.GATSBY_EXPERIMENTAL_LMDB_STORE": `true`,
         "process.env.GATSBY_SKIP_WRITING_SCHEMA_TO_FILE": `true`,
         SCHEMA_SNAPSHOT: JSON.stringify(schemaSnapshotString),
+        "process.env.GATSBY_LOGGER": JSON.stringify(`yurnalist`),
       }),
       process.env.GATSBY_WEBPACK_LOGGING?.includes(`query-engine`) &&
         new WebpackLoggingPlugin(rootDir, reporter, isVerbose),