+++ packages/gatsby/src/utils/api-node-docs.ts
@@ -299,11 +299,8 @@ export const createSchemaCustomization = true
  *   built schema from `info.schema`.
  * * Gatsby's data layer, including all internal query capabilities, is
  *   exposed on [`context.nodeModel`](/docs/node-model/). The node store can be
- *   queried directly with `getAllNodes`, `getNodeById` and `getNodesByIds`,
- *   while more advanced queries can be composed with `runQuery`. Note that
- *   `runQuery` will call field resolvers before querying, so e.g. foreign-key
- *   fields will be expanded to full nodes. The other methods on `nodeModel`
- *   don't do this.
+ *   queried directly with `findOne`, `getNodeById` and `getNodesByIds`,
+ *   while more advanced queries can be composed with `findAll`.
  * * It is possible to add fields to the root `Query` type.
  * * When using the first resolver argument (`source` in the example below,
  *   often also called `parent` or `root`), take care of the fact that field
@@ -334,12 +331,11 @@ export const createSchemaCustomization = true
  *     Query: {
  *       allRecentPosts: {
  *         type: [`BlogPost`],
- *         resolve: (source, args, context, info) => {
- *           const posts = context.nodeModel.getAllNodes({ type: `BlogPost` })
- *           const recentPosts = posts.filter(
+ *         resolve: async (source, args, context, info) => {
+ *           const { entries } = await context.nodeModel.findAll({ type: `BlogPost` })
+ *           return entries.filter(
  *             post => post.publishedAt > Date.UTC(2018, 0, 1)
  *           )
- *           return recentPosts
  *         }
  *       }
  *     }